Issue 892: refactored everything reusable out of school data loader test into base parent class.
