+++ src/Ilios/CoreBundle/Tests/DataFixtures/ORM/LoadSchoolDataTest.php
@@ -2,76 +2,15 @@
 
 namespace Ilios\CoreBundle\Tests\DataFixtures\ORM;
 
-use Ilios\CoreBundle\Entity\Manager\ManagerInterface;
-use Ilios\CoreBundle\Entity\Manager\SchoolManager;
-
 use Ilios\CoreBundle\Entity\Manager\SchoolManagerInterface;
 use Ilios\CoreBundle\Entity\SchoolInterface;
-
-use Liip\FunctionalTestBundle\Test\WebTestCase;
-
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Class LoadSchoolDataTest
  * @package Ilios\CoreBundle\Tests\DataFixtures\ORM
  */
-class LoadSchoolDataTest extends WebTestCase
+class LoadSchoolDataTest extends AbstractDataFixtureTest
 {
-    /**
-     * @var ContainerInterface
-     */
-    protected $container;
-
-    /**
-     * @var resource
-     */
-    protected $dataFile;
-
-    /**
-     * @var ManagerInterface
-     */
-    protected $em;
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setUp()
-    {
-        $this->container = static::createClient()->getContainer();
-        $this->loadFixtures($this->getFixtures());
-        $this->loadDataFile($this->getDataFileName());
-        $this->loadEntityManager($this->getEntityManagerServiceKey());
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function tearDown(){
-        fclose($this->dataFile);
-    }
-
-    /**
-     * @param $fileName
-     */
-    protected function loadDataFile($fileName)
-    {
-        /**
-         * @var FileLocator $fileLocator
-         */
-        $fileLocator = $this->container->get('file_locator');
-        $path = $fileLocator->locate('@IliosCoreBundle/Resources/dataimport/' . basename($fileName));
-        $this->dataFile = fopen($path, 'r');
-    }
-
-    /**
-     * @param $serviceKey
-     */
-    protected function loadEntityManager($serviceKey)
-    {
-        $this->em = $this->container->get($serviceKey);
-    }
-
     /**
      * @return string
      */
@@ -99,55 +38,31 @@ class LoadSchoolDataTest extends WebTestCase
     }
 
     /**
-     * @covers Ilios\CoreBundle\DataFixtures\ORM\LoadSchoolData::load
+     * @param array $data
+     * @param SchoolInterface $entity
      */
-    public function testLoad()
+    protected function assertDataEquals(array $data, $entity)
     {
-        /**
-         * @var SchoolManager $em
-         */
-        $em = $this->container->get('ilioscore.school.manager');
-
-        $first = true;
-        while (($data = fgetcsv($this->dataFile)) !== FALSE) {
-            // step over the first row
-            // since it contains the field names
-            if ($first) {
-                $first = false;
-                continue;
-            }
-            $entity = $this->getEntityForRow($data);
-            $this->assertDataIntegrity($data, $entity);
-        }
-
+        // `school_id`,`template_prefix`,`title`,`ilios_administrator_email`,`deleted`,`change_alert_recipients`
+        $this->assertEquals($data[0], $entity->getId());
+        $this->assertEquals($data[1], $entity->getTemplatePrefix());
+        $this->assertEquals($data[2], $entity->getTitle());
+        $this->assertEquals($data[3], $entity->getIliosAdministratorEmail());
+        $this->assertEquals((boolean) $data[4], $entity->isDeleted());
+        $this->assertEquals($data[5], $entity->getChangeAlertRecipients());
     }
 
     /**
-     * @param array $row
+     * @param array $data
      * @return SchoolInterface
+     * @override
      */
-    protected function getEntityForRow (array $row)
+    protected function getEntity(array $data)
     {
         /**
          * @var SchoolManagerInterface $em
          */
         $em = $this->em;
-        return $em->findSchoolBy(['id' => $row[0]]);
-
-    }
-
-    /**
-     * @param array $row
-     * @param SchoolInterface $entity
-     */
-    protected function assertDataIntegrity(array $row, $entity)
-    {
-        // `school_id`,`template_prefix`,`title`,`ilios_administrator_email`,`deleted`,`change_alert_recipients`
-        $this->assertEquals($row[0], $entity->getId());
-        $this->assertEquals($row[1], $entity->getTemplatePrefix());
-        $this->assertEquals($row[2], $entity->getTitle());
-        $this->assertEquals($row[3], $entity->getIliosAdministratorEmail());
-        $this->assertEquals((boolean) $row[4], $entity->isDeleted());
-        $this->assertEquals($row[5], $entity->getChangeAlertRecipients());
+        return $em->findSchoolBy(['id' => $data[0]]);
     }
 }
